Separated GQ_Characterize declaration and implementation
